+++ popacket-service/src/main/java/popacketservice/popacketservice/service/ShipmentService.java
@@ -11,7 +11,6 @@ import popacketservice.popacketservice.model.entity.Shipment;
 import popacketservice.popacketservice.repository.ShipmentRepository;
 import popacketservice.popacketservice.repository.ShippingRateRepository;
 
-import java.math.BigDecimal;
 
 @Service
 @Data
@@ -27,27 +26,11 @@ public class ShipmentService {
     @Autowired
     private ShipmentMapper shipmentMapper;
 
-    public ShipmentResponseDTO cancelShipmentById(Long id) {
-        Shipment shipmentTemp = shipmentRepository.getShipmentById(id).orElseThrow();
-        shipmentTemp.setStatus("cancelado");
-        shipmentRepository.save(shipmentTemp);
-        return shipmentMapper.convertToDTO(shipmentTemp);
-    }
-
-    public Double getShipmentCost(Double weight, String serviceType) {
-        BigDecimal priceBase = shipmentRateRepository.getBasePrice(BigDecimal.valueOf(weight), serviceType);
-        BigDecimal pricePerKilometer = shipmentRateRepository.getPricePerKilometer(BigDecimal.valueOf(weight), serviceType);
-        Double price = priceBase.add(pricePerKilometer).doubleValue();
-        return price;
-    }
-
-
     public ShipmentResponseDTO getShipmentById(Long id) {
         Shipment shipmentTemp = shipmentRepository.getShipmentById(id).orElseThrow();
         //Object[] shipmentStatus = shipmentRepository.getStatusShipmentById(id).orElseThrow();
         return shipmentMapper.convertToDTO(shipmentTemp);
     }
-
 
     public Object[] getStatusShipmentById(Long id) {
         //Shipment shipmentTemp = shipmentRepository.getShipmentById(id).orElseThrow();
