--- popacket-service/src/test/java/popacketservice/popacketservice/ShipmentServiceTest.java
@@ -1,126 +1,0 @@
-package popacketservice.popacketservice;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import popacketservice.popacketservice.mapper.ShipmentMapper;
-import popacketservice.popacketservice.model.dto.ShipmentResponseDTO;
-import popacketservice.popacketservice.model.entity.Shipment;
-import popacketservice.popacketservice.repository.ShipmentRepository;
-import popacketservice.popacketservice.repository.ShippingRateRepository;
-import popacketservice.popacketservice.service.ShipmentService;
-
-import java.math.BigDecimal;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.Mockito.*;
-
-public class ShipmentServiceTest {
-
-    @Mock
-    private ShipmentRepository shipmentRepository;
-
-    @Mock
-    private ShippingRateRepository shipmentRateRepository;
-
-    @Mock
-    private ShipmentMapper shipmentMapper;
-
-    @InjectMocks
-    private ShipmentService shipmentService;
-
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    public void getStatusShipmentById_Success() {
-
-        Long shipmentId = 1L;
-        Object[] expectedShipmentStatus = {"En camino", "2024-06-10"};
-
-        when(shipmentRepository.getStatusShipmentByIdOb(anyLong())).thenReturn(Optional.of(expectedShipmentStatus));
-
-
-        Object[] result = shipmentService.getStatusShipmentById(shipmentId);
-
-
-        assertArrayEquals(expectedShipmentStatus, result);
-        verify(shipmentRepository, times(1)).getStatusShipmentByIdOb(shipmentId);
-    }
-
-    @Test
-    public void getStatusShipmentById_NotFound() {
-
-        Long shipmentId = 1L;
-
-        when(shipmentRepository.getStatusShipmentByIdOb(anyLong())).thenReturn(Optional.empty());
-
-
-        RuntimeException exception = assertThrows(RuntimeException.class, () -> {
-            shipmentService.getStatusShipmentById(shipmentId);
-        });
-
-        assertEquals("El id del Envio no existe", exception.getMessage());
-        verify(shipmentRepository, times(1)).getStatusShipmentByIdOb(shipmentId);
-    }
-    @Test
-    public void testCancelShipmentById() {
-        // Datos de prueba
-        Long shipmentId = 1L;
-        Shipment shipment = new Shipment();
-        shipment.setId(shipmentId);
-        shipment.setStatus("activo");
-
-        ShipmentResponseDTO shipmentResponseDTO = new ShipmentResponseDTO();
-        shipmentResponseDTO.setId(shipmentId);
-        shipmentResponseDTO.setStatus("cancelado");
-
-        // Simulación del comportamiento
-        when(shipmentRepository.getShipmentById(shipmentId)).thenReturn(Optional.of(shipment));
-        when(shipmentMapper.convertToDTO(any(Shipment.class))).thenReturn(shipmentResponseDTO);
-
-        // Llamada al método a probar
-        ShipmentResponseDTO result = shipmentService.cancelShipmentById(shipmentId);
-
-        // Verificación
-        assertEquals("cancelado", result.getStatus());
-        verify(shipmentRepository, times(1)).save(any(Shipment.class));
-    }
-
-    @Test
-    public void testCancelShipmentById_NotFound() {
-        // Datos de prueba
-        Long shipmentId = 1L;
-
-        // Simulación del comportamiento
-        when(shipmentRepository.getShipmentById(shipmentId)).thenReturn(Optional.empty());
-
-        // Llamada al método a probar y verificación de la excepción
-        assertThrows(RuntimeException.class, () -> shipmentService.cancelShipmentById(shipmentId));
-    }
-
-    @Test
-    public void testGetShipmentCost() {
-        // Datos de prueba
-        Double weight = 10.0;
-        String serviceType = "Economico";
-        BigDecimal basePrice = BigDecimal.valueOf(50);
-        BigDecimal pricePerKilometer = BigDecimal.valueOf(5);
-
-        // Simulación del comportamiento
-        when(shipmentRateRepository.getBasePrice(BigDecimal.valueOf(weight), serviceType)).thenReturn(basePrice);
-        when(shipmentRateRepository.getPricePerKilometer(BigDecimal.valueOf(weight), serviceType)).thenReturn(pricePerKilometer);
-
-        // Llamada al método a probar
-        Double cost = shipmentService.getShipmentCost(weight, serviceType);
-
-        // Verificación
-        assertEquals(55.0, cost);
-    }
-}
